+++ crates/core/src/database/common.rs
@@ -38,7 +38,7 @@ macro_rules! auto_method {
                 Err(e) => return Err(Error::DatabaseConnection(e.to_string())),
             };
 
-            let res = query_row!(&conn, $query, &[&(id as isize)], |x| {
+            let res = query_row!(&conn, $query, &[&(id as i64)], |x| {
                 Ok(Self::$select_fn(x))
             });
 
@@ -61,7 +61,7 @@ macro_rules! auto_method {
                 Err(e) => return Err(Error::DatabaseConnection(e.to_string())),
             };
 
-            let res = query_row!(&conn, $query, &[&(id as isize)], |x| {
+            let res = query_row!(&conn, $query, &[&(id as i64)], |x| {
                 Ok(Self::$select_fn(x))
             });
